+++ Cargo.toml
@@ -49,13 +49,13 @@ mockall = { version = "0.11", optional = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tracerr = "0.3"
-url = "2.2"
+url = "2.3"
 
 [target.'cfg(target_os = "android")'.dependencies]
 android_logger = "0.11"
 
 [target.'cfg(any(target_os = "windows", target_os = "linux", target_os = "macos"))'.dependencies]
-simple_logger = "2.1"
+simple_logger = "2.3"
 
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
 dart-sys = "2.0"
