+++ flutter/lib/src/native/connection_handle.dart
@@ -2,7 +2,7 @@ import 'dart:ffi';
 
 import '../interface/connection_handle.dart';
 import '../interface/remote_media_track.dart';
-import '../interface/track_kinds.dart';
+import '../interface/track.dart';
 import '../util/move_semantic.dart';
 import '/src/util/rust_handles_storage.dart';
 import 'ffi/foreign_value.dart';
