+++ flutter/lib/src/web/remote_media_track.dart
@@ -3,7 +3,7 @@ import 'package:medea_flutter_webrtc/medea_flutter_webrtc.dart' as webrtc;
 import 'package:medea_flutter_webrtc/src/platform/web/media_stream_track.dart';
 
 import '../interface/remote_media_track.dart';
-import '../interface/track_kinds.dart';
+import '../interface/track.dart';
 import '../util/move_semantic.dart';
 import 'exceptions.dart';
 import 'jason_wasm.dart' as wasm;
